updated schematic/board with correct VDD wiring
Might not be in sync now (between schematic and board), but the data is
correct. The problem was the battery output and switch output
accidentally had the same label, so the switch was being bypassed.
Correct this routing, as well as routing VDD to \bar{RESET} so we
wouldn't have to lose that functionality by changing RESET to a GPIO